requirements.txt
absl-py==0.9.0
gensim==3.8.1
numpy==1.18.1
PyYAML==5.3.1
scikit-learn==0.22.1
scipy==1.4.1
tensorflow-determinism==0.3.0
tensorflow-gpu==2.1.0
tqdm==4.43.0
